Add requirements.txt
Whilst setup.py lists the required python modules it allows attempted
installation of the latest versions of each.

Unfortunately there is at least the problem that simplejson's latest
version attempts to install some dependencies which won't install on
Python 2.7.

Thus we nail down the versions of all the required modules as per
Athanasius' experience of what actually works on a Debian system.