setup.py: Fix permissions on monitor/schemas after copy
In Python 2.7 shutil.copytree() *always* preserves the file mode.  I
don't want to rely on changing them in the source as that's fragile.

So, write our own recursive function to fix them after the copy.

Once we move to Python 3 we can use the copy_function argument, combined
with the umask we already set.