docs/Developers: Allow for `""` gameversion/build values.
The emphasis here is on strongly encouraging Senders to get the actual, full
and correct, values to Listeners.  But where they can't, be that due to
data source limitations (CAPI), or their code architecture (common EDDN
sending code that has no concept of the data source), they're allowed to
send a less than optimal value.

**BUT**, they still should start sending the fields as outlined.